// src/app/list-offers/models/offer.ts
export interface Offer 
{
    title: string
    salePrice: string
    normalPrice: string
    thumb: string
    discount?: string
    dealID: string
    dealRating: string
    gameID: string
    internalName: string
    isOnSale: string
    lastChange: string
    metacriticLink: string
    metacriticScore: string
    releaseDate: string
    savings: string
    steamAppID: string
    steamRatingCount: string
    steamRatingPercent: string
    steamRatingText: string
    storeID: string
}